# Conflicts:
#	Assets/Scripts/DeathState.cs.meta
#	Assets/Scripts/states/ChargedAttackState.cs
#	Assets/prefabs/Player.prefab
#	Assets/prefabs/player2.prefab

+++ Assets/Scripts/HealthBarUI.cs
@@ -14,7 +14,7 @@ public class HealthBarUI : MonoBehaviour
     {
         if (target == null || fillBar == null) return;
 
-        int currentHP = Mathf.Clamp(target.GetCurrentHP(), 0, target.maxHP);
+        float currentHP = Mathf.Clamp(target.GetCurrentHP(), 0, target.maxHP);
         float ratio = (float)currentHP / target.maxHP;
 
         fillBar.fillAmount = ratio;
